Улучшения по стилю кодирования
- В стиле кодирования от Microsoft для С++ классы имеют префикс C, а их файлы - нет
- Применил constructor initializer lists в конструкторе CBankClient
- Убрал циклический #include между Bank.h и BankClient.h, который не будет работать в рамках препроцессора и раздельной компиляции C/C++
