+++ PP2/Bank.h
@@ -1,8 +1,8 @@
 #pragma once
 #include <iostream>
 #include <vector>
-#include "CBankClient.h"
-class CBankClient;
+
+#include "BankClient.h"
 
 class CBank
 {
@@ -10,6 +10,7 @@ public:
 	CBank();
 	CBankClient* CreateClient();
 	void UpdateClientBalance(CBankClient& client, int value);
+
 private:
 	std::vector<CBankClient> m_clients;
 	int m_totalBalance;
