+++ PP2/BankClient.h
@@ -1,0 +1,21 @@
+#pragma once
+#include <Windows.h>
+
+class CBank;
+
+class CBankClient
+{
+public:
+	unsigned GetId();
+
+private:
+	CBankClient(CBank *bank, unsigned id);
+	static unsigned GetSleepDuration(CBankClient *client);
+	static unsigned GetBalanceChangeValue();
+	static DWORD WINAPI ThreadFunction(LPVOID lpParam);
+
+	friend CBank;
+
+	CBank *m_bank;
+	unsigned m_id;
+};
