+++ PP2/PP2.vcxproj
@@ -145,14 +145,14 @@
     <Text Include="ReadMe.txt" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="CBank.h" />
-    <ClInclude Include="CBankClient.h" />
+    <ClInclude Include="Bank.h" />
+    <ClInclude Include="BankClient.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="CBank.cpp" />
-    <ClCompile Include="CBankClient.cpp" />
+    <ClCompile Include="Bank.cpp" />
+    <ClCompile Include="BankClient.cpp" />
     <ClCompile Include="PP2.cpp" />
     <ClCompile Include="stdafx.cpp" />
   </ItemGroup>
